[tests] Move StubBaseVirtualMachine to the .cpp.
There was one header file containing only this definition. We moved it
to the .cpp where it's used because it's useless outside it.
